+++ src/mainwindow.h
@@ -65,7 +65,6 @@ signals:
 private slots:
     void on_translateButton_clicked();
     void on_swapButton_clicked();
-    void on_copyToSourceButton_clicked();
     void on_settingsButton_clicked();
 
     void on_playSourceButton_clicked();